pyproject.toml
[tool.poetry]
name = "datawagon"
version = "0.1.0"
description = ""
authors = ["xh4vm <h4vm@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
black = "^23.10.1"
pydantic = "^2.4.2"
fastapi = "^0.104.1"
orjson = "^3.9.10"
backoff = "^2.2.1"
jinja2 = "^3.1.2"
clickhouse-driver = "^0.2.6"
aiohttp = "^3.8.6"
dependency-injector = "^4.41.0"
gunicorn = "^21.2.0"
uvicorn = "^0.23.2"
aiokafka = "^0.8.1"
starlette-context = "^0.3.6"
pydantic-settings = "^2.0.3"
loguru = "^0.7.2"
osmium = "^3.6.0"
shapely = "^2.0.2"
geopandas = "^0.14.0"
geojson = "^3.1.0"
celery = "^5.3.4"
psycopg2-binary = "^2.9.9"
django = "^4.2.7"
django-rest-framework = "^0.1.0"
enforce-typing = "^1.0.0.post1"
split-settings = "^1.0.0"
python-dotenv = "^1.0.0"
apache-airflow-providers-celery = "^3.4.1"
django-debug-toolbar = "^4.2.0"
django-split-settings = "^1.2.0"
django-cors-headers = "^4.3.0"
sh = "^2.0.6"
wget = "^3.2"


[tool.poetry.group.build.dependencies]
django = "^4.2.7"
django-debug-toolbar = "^4.2.0"
django-split-settings = "^1.2.0"
django-cors-headers = "^4.3.0"
django-rest-framework = "^0.1.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
osmium = "^3.6.0"
shapely = "^2.0.2"
geopandas = "^0.14.0"
geojson = "^3.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
